detector/requirements.txt
absl-py==1.4.0
aiogram==2.25.1
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.1
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.7.*
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
protobuf==3.19.*
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
pyshark==0.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tensorflow-cpu==2.7.0
termcolor==2.2.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
